Enabled email tests as well + added required fields in app.yaml spec
